Move TokenService project into App project as this really is too simple for such things. Upgrade project ot VS2013.
