+++ .gitignore
@@ -134,6 +134,7 @@ publish/
 *.pubxml
 *.publishproj
 
+**/.nuget/NuGet.exe
 # NuGet Packages
 *.nupkg
 # The packages folder can be ignored because of Package Restore
