+++ src/App/Forms/MainWindow.resx
