+++ src/App/Model/DecodedToken.cs
@@ -1,10 +1,10 @@
-﻿namespace AzureTokenMaker.Service {
+﻿namespace AzureTokenMaker.App.Model {
     public sealed class DecodedToken {
         private readonly string _signature;
         private readonly string _claims;
         private readonly string _header;
 
-        internal DecodedToken(string signature, string claims, string header) {
+        public DecodedToken(string signature, string claims, string header) {
             _signature = signature;
             _claims = claims;
             _header = header;
