+++ src/App/Model/ProfileData.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using AzureTokenMaker.App.Services;
+
+namespace AzureTokenMaker.App.Model {
+    public sealed class ProfileData
+    {
+
+    }
+}
