--- src/App/ProfileData.cs
@@ -1,8 +1,0 @@
-﻿using System;
-using AzureTokenMaker.Service;
-
-namespace AzureTokenMaker.App {
-    sealed class ProfileData {
-
-    }
-}
