+++ src/AzureTokenMaker.sln
@@ -1,9 +1,16 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AzureTokenMaker.Service", "src\TokenService\AzureTokenMaker.Service.csproj", "{3C5E3DD6-5A4D-41CD-A6B9-EB14F8D50536}"
+# Visual Studio 2013
+VisualStudioVersion = 12.0.40629.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AzureTokenMaker.App", "App\AzureTokenMaker.App.csproj", "{5F082684-0E44-491A-9AD6-EF37F7B440F5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AzureTokenMaker.App", "src\App\AzureTokenMaker.App.csproj", "{5F082684-0E44-491A-9AD6-EF37F7B440F5}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{F71A66A8-4E22-4059-9D20-158A563EF5E5}"
+	ProjectSection(SolutionItems) = preProject
+		.nuget\NuGet.Config = .nuget\NuGet.Config
+		.nuget\NuGet.exe = .nuget\NuGet.exe
+		.nuget\NuGet.targets = .nuget\NuGet.targets
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +18,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3C5E3DD6-5A4D-41CD-A6B9-EB14F8D50536}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3C5E3DD6-5A4D-41CD-A6B9-EB14F8D50536}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3C5E3DD6-5A4D-41CD-A6B9-EB14F8D50536}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3C5E3DD6-5A4D-41CD-A6B9-EB14F8D50536}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5F082684-0E44-491A-9AD6-EF37F7B440F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5F082684-0E44-491A-9AD6-EF37F7B440F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5F082684-0E44-491A-9AD6-EF37F7B440F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
